requirements.txt
Flask==2.2.3
pandas==1.5.3
pydantic==1.10.7
python-dateutil==2.8.2
requests==2.28.2
selenium==4.8.3